MVP Starting point:
Well this commit is inteded to be a good starting point onwards to the MVP's defined end date.

This has an initial version/proposal of structure for the 3 tickets we selected as part of the MVP.

So the MainContainer is already existing, we have also a router, we have some template elements in order to be able to test correctly the router (Album, playlist, profile, Queue, etc) and there are also some parts of the HomePage and MusicPlayer components.